requirements.txt
altgraph==0.17
fbs==0.8.6
future==0.18.2
macholib==1.14
pefile==2019.4.18
PyInstaller==3.4
PySide2==5.14.2.1
pyusb==1.0.2
shiboken2==5.14.2.1
